+++ R/auto_app_bsky_unspecced.R
@@ -3,7 +3,7 @@
 #' @noRd
 app_bsky_unspecced_get_config <- function(.token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.getConfig",
+    name = "bsky.social/xrpc/app.bsky.unspecced.getConfig",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -19,7 +19,103 @@ app_bsky_unspecced_get_config <- function(.token = NULL, .return = c("json", "re
 #' @noRd
 app_bsky_unspecced_get_popular_feed_generators <- function(limit = NULL, cursor = NULL, query = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.getPopularFeedGenerators",
+    name = "bsky.social/xrpc/app.bsky.unspecced.getPopularFeedGenerators",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_feeds
+#' Get a list of suggested feeds
+#' @noRd
+app_bsky_unspecced_get_suggested_feeds <- function(limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedFeeds",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_feeds_skeleton
+#' Get a skeleton of suggested feeds. Intended to be called and hydrated by app.bsky.unspecced.getSuggestedFeeds
+#' @noRd
+app_bsky_unspecced_get_suggested_feeds_skeleton <- function(viewer = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedFeedsSkeleton",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_starter_packs
+#' Get a list of suggested starterpacks
+#' @noRd
+app_bsky_unspecced_get_suggested_starter_packs <- function(limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedStarterPacks",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_starter_packs_skeleton
+#' Get a skeleton of suggested starterpacks. Intended to be called and hydrated by app.bsky.unspecced.getSuggestedStarterpacks
+#' @noRd
+app_bsky_unspecced_get_suggested_starter_packs_skeleton <- function(viewer = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedStarterPacksSkeleton",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_users
+#' Get a list of suggested users
+#' @noRd
+app_bsky_unspecced_get_suggested_users <- function(category = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedUsers",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_suggested_users_skeleton
+#' Get a skeleton of suggested users. Intended to be called and hydrated by app.bsky.unspecced.getSuggestedUsers
+#' @noRd
+app_bsky_unspecced_get_suggested_users_skeleton <- function(viewer = NULL, category = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestedUsersSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -35,7 +131,7 @@ app_bsky_unspecced_get_popular_feed_generators <- function(limit = NULL, cursor 
 #' @noRd
 app_bsky_unspecced_get_suggestions_skeleton <- function(viewer = NULL, limit = NULL, cursor = NULL, relativeToDid = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.getSuggestionsSkeleton",
+    name = "bsky.social/xrpc/app.bsky.unspecced.getSuggestionsSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -51,7 +147,55 @@ app_bsky_unspecced_get_suggestions_skeleton <- function(viewer = NULL, limit = N
 #' @noRd
 app_bsky_unspecced_get_tagged_suggestions <- function(.token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.getTaggedSuggestions",
+    name = "bsky.social/xrpc/app.bsky.unspecced.getTaggedSuggestions",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_trending_topics
+#' Get a list of trending topics
+#' @noRd
+app_bsky_unspecced_get_trending_topics <- function(viewer = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getTrendingTopics",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_trends
+#' Get the current trends on the network
+#' @noRd
+app_bsky_unspecced_get_trends <- function(limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getTrends",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' app_bsky_unspecced_get_trends_skeleton
+#' Get the skeleton of trends on the network. Intended to be called and then hydrated through app.bsky.unspecced.getTrends
+#' @noRd
+app_bsky_unspecced_get_trends_skeleton <- function(viewer = NULL, limit = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/app.bsky.unspecced.getTrendsSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -67,7 +211,7 @@ app_bsky_unspecced_get_tagged_suggestions <- function(.token = NULL, .return = c
 #' @noRd
 app_bsky_unspecced_search_actors_skeleton <- function(q, viewer = NULL, typeahead = NULL, limit = NULL, cursor = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.searchActorsSkeleton",
+    name = "bsky.social/xrpc/app.bsky.unspecced.searchActorsSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -83,7 +227,7 @@ app_bsky_unspecced_search_actors_skeleton <- function(q, viewer = NULL, typeahea
 #' @noRd
 app_bsky_unspecced_search_posts_skeleton <- function(q, sort = NULL, since = NULL, until = NULL, mentions = NULL, author = NULL, lang = NULL, domain = NULL, url = NULL, tag = NULL, viewer = NULL, limit = NULL, cursor = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.searchPostsSkeleton",
+    name = "bsky.social/xrpc/app.bsky.unspecced.searchPostsSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -99,7 +243,7 @@ app_bsky_unspecced_search_posts_skeleton <- function(q, sort = NULL, since = NUL
 #' @noRd
 app_bsky_unspecced_search_starter_packs_skeleton <- function(q, viewer = NULL, limit = NULL, cursor = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/app.bsky.unspecced.searchStarterPacksSkeleton",
+    name = "bsky.social/xrpc/app.bsky.unspecced.searchStarterPacksSkeleton",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())