+++ R/auto_chat_bsky_convo.R
@@ -1,9 +1,41 @@
+#' chat_bsky_convo_accept_convo
+#' 
+#' @noRd
+chat_bsky_convo_accept_convo <- function(convoId, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/chat.bsky.convo.acceptConvo",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "POST"
+  )
+}
+
+
+
+#' chat_bsky_convo_add_reaction
+#' Adds an emoji reaction to a message. Requires authentication. It is idempotent, so multiple calls from the same user with the same emoji result in a single reaction.
+#' @noRd
+chat_bsky_convo_add_reaction <- function(convoId, messageId, value, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/chat.bsky.convo.addReaction",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "POST"
+  )
+}
+
+
+
 #' chat_bsky_convo_delete_message_for_self
 #' 
 #' @noRd
 chat_bsky_convo_delete_message_for_self <- function(convoId, messageId, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.deleteMessageForSelf",
+    name = "bsky.social/xrpc/chat.bsky.convo.deleteMessageForSelf",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -19,7 +51,23 @@ chat_bsky_convo_delete_message_for_self <- function(convoId, messageId, .token =
 #' @noRd
 chat_bsky_convo_get_convo <- function(convoId, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.getConvo",
+    name = "bsky.social/xrpc/chat.bsky.convo.getConvo",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "GET"
+  )
+}
+
+
+
+#' chat_bsky_convo_get_convo_availability
+#' Get whether the requester and the other members can chat. If an existing convo is found for these members, it is returned.
+#' @noRd
+chat_bsky_convo_get_convo_availability <- function(members, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/chat.bsky.convo.getConvoAvailability",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -35,7 +83,7 @@ chat_bsky_convo_get_convo <- function(convoId, .token = NULL, .return = c("json"
 #' @noRd
 chat_bsky_convo_get_convo_for_members <- function(members, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.getConvoForMembers",
+    name = "bsky.social/xrpc/chat.bsky.convo.getConvoForMembers",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -51,7 +99,7 @@ chat_bsky_convo_get_convo_for_members <- function(members, .token = NULL, .retur
 #' @noRd
 chat_bsky_convo_get_log <- function(cursor = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.getLog",
+    name = "bsky.social/xrpc/chat.bsky.convo.getLog",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -67,7 +115,7 @@ chat_bsky_convo_get_log <- function(cursor = NULL, .token = NULL, .return = c("j
 #' @noRd
 chat_bsky_convo_get_messages <- function(convoId, limit = NULL, cursor = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.getMessages",
+    name = "bsky.social/xrpc/chat.bsky.convo.getMessages",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -83,7 +131,7 @@ chat_bsky_convo_get_messages <- function(convoId, limit = NULL, cursor = NULL, .
 #' @noRd
 chat_bsky_convo_leave_convo <- function(convoId, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.leaveConvo",
+    name = "bsky.social/xrpc/chat.bsky.convo.leaveConvo",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -97,9 +145,9 @@ chat_bsky_convo_leave_convo <- function(convoId, .token = NULL, .return = c("jso
 #' chat_bsky_convo_list_convos
 #' 
 #' @noRd
-chat_bsky_convo_list_convos <- function(limit = NULL, cursor = NULL, .token = NULL, .return = c("json", "resp")) {
+chat_bsky_convo_list_convos <- function(limit = NULL, cursor = NULL, readState = NULL, status = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.listConvos",
+    name = "bsky.social/xrpc/chat.bsky.convo.listConvos",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -115,7 +163,23 @@ chat_bsky_convo_list_convos <- function(limit = NULL, cursor = NULL, .token = NU
 #' @noRd
 chat_bsky_convo_mute_convo <- function(convoId, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.muteConvo",
+    name = "bsky.social/xrpc/chat.bsky.convo.muteConvo",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "POST"
+  )
+}
+
+
+
+#' chat_bsky_convo_remove_reaction
+#' Removes an emoji reaction from a message. Requires authentication. It is idempotent, so multiple calls from the same user with the same emoji result in that reaction not being present, even if it already wasn't.
+#' @noRd
+chat_bsky_convo_remove_reaction <- function(convoId, messageId, value, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/chat.bsky.convo.removeReaction",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -131,7 +195,7 @@ chat_bsky_convo_mute_convo <- function(convoId, .token = NULL, .return = c("json
 #' @noRd
 chat_bsky_convo_send_message <- function(convoId, message, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.sendMessage",
+    name = "bsky.social/xrpc/chat.bsky.convo.sendMessage",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -147,7 +211,7 @@ chat_bsky_convo_send_message <- function(convoId, message, .token = NULL, .retur
 #' @noRd
 chat_bsky_convo_send_message_batch <- function(items, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.sendMessageBatch",
+    name = "bsky.social/xrpc/chat.bsky.convo.sendMessageBatch",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -163,7 +227,23 @@ chat_bsky_convo_send_message_batch <- function(items, .token = NULL, .return = c
 #' @noRd
 chat_bsky_convo_unmute_convo <- function(convoId, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.unmuteConvo",
+    name = "bsky.social/xrpc/chat.bsky.convo.unmuteConvo",
+    params = as.list(match.call())[-1] |>
+      purrr::imap(~ {
+        eval(.x, envir = parent.frame())
+      }),
+    req_method = "POST"
+  )
+}
+
+
+
+#' chat_bsky_convo_update_all_read
+#' 
+#' @noRd
+chat_bsky_convo_update_all_read <- function(status = NULL, .token = NULL, .return = c("json", "resp")) {
+  make_request(
+    name = "bsky.social/xrpc/chat.bsky.convo.updateAllRead",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())
@@ -179,7 +259,7 @@ chat_bsky_convo_unmute_convo <- function(convoId, .token = NULL, .return = c("js
 #' @noRd
 chat_bsky_convo_update_read <- function(convoId, messageId = NULL, .token = NULL, .return = c("json", "resp")) {
   make_request(
-    hostname = "bsky.social/xrpc/chat.bsky.convo.updateRead",
+    name = "bsky.social/xrpc/chat.bsky.convo.updateRead",
     params = as.list(match.call())[-1] |>
       purrr::imap(~ {
         eval(.x, envir = parent.frame())