+++ R/lists.r
@@ -3,6 +3,7 @@
 #' Get all info about a starter pack and the users on it.
 #'
 #' @param starter_pack the URL of a starter pack
+#' @param actor user handle of account to query for starter packs.
 #' @inheritParams search_user
 #'
 #' @returns a data frame of users and list info
@@ -21,6 +22,41 @@ get_starter_pack <- function(starter_pack,
   if (!parse) return(res)
   return(parse_starter_packs(res))
 }
+
+
+#' @export
+#' @rdname get_starter_pack
+get_actor_starter_packs <- function(actor,
+                                    limit = NULL,
+                                    cursor = NULL,
+                                    .token = NULL) {
+  resp <- list(cursor = cursor %||% "")
+  res <- list()
+  while (purrr::pluck_exists(resp, "cursor")) {
+    resp <- do.call(
+      app_bsky_graph_get_actor_starter_packs,
+      list(
+        actor = actor,
+        limit = limit,
+        cursor = purrr::pluck(resp, "cursor"),
+        .token = .token
+      )
+    )
+    res <- c(res, resp)
+  }
+
+  if (!parse) {
+    return(res)
+  }
+  out <- res |>
+    unlist(recursive = FALSE) |>
+    purrr::map(as_tibble_onerow) |>
+    dplyr::bind_rows()
+
+  attr(out, "cursor") <- purrr::pluck(res, "cursor")
+  return(out)
+}
+
 
 
 #' Get List